ai-brand-automator/requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0
freezegun==1.4.0
responses==0.24.1
hypothesis==6.92.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3
bandit==1.7.6

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Debugging
ipython==8.19.0
ipdb==0.13.13

# Performance
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Type Checking
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-requests==2.31.0.20231231
